genlock: Backport from 4.1.H.0.4 sources

+++ drivers/base/genlock.c
@@ -1,4 +1,5 @@
 /* Copyright (c) 2011-2012, Code Aurora Forum. All rights reserved.
+ * Copyright (C) 2011 Sony Ericsson Mobile Communications AB.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2 and
@@ -289,7 +290,7 @@ static int _genlock_lock(struct genlock *lock, struct genlock_handle *handle,
 {
 	unsigned long irqflags;
 	int ret = 0;
-	unsigned long ticks = msecs_to_jiffies(timeout);
+	unsigned int ticks = msecs_to_jiffies(timeout);
 
 	spin_lock_irqsave(&lock->lock, irqflags);
 
@@ -360,7 +361,7 @@ static int _genlock_lock(struct genlock *lock, struct genlock_handle *handle,
 	/* Wait while the lock remains in an incompatible state */
 
 	while (lock->state != _UNLOCKED) {
-		signed long elapsed;
+		unsigned int elapsed;
 
 		spin_unlock_irqrestore(&lock->lock, irqflags);
 
@@ -374,7 +375,7 @@ static int _genlock_lock(struct genlock *lock, struct genlock_handle *handle,
 			goto done;
 		}
 
-		ticks = (unsigned long) elapsed;
+		ticks = elapsed;
 	}
 
 dolock:
@@ -448,7 +449,7 @@ int genlock_wait(struct genlock_handle *handle, uint32_t timeout)
 	struct genlock *lock;
 	unsigned long irqflags;
 	int ret = 0;
-	unsigned long ticks = msecs_to_jiffies(timeout);
+	unsigned int ticks = msecs_to_jiffies(timeout);
 
 	if (IS_ERR_OR_NULL(handle)) {
 		GENLOCK_LOG_ERR("Invalid handle\n");
@@ -475,7 +476,7 @@ int genlock_wait(struct genlock_handle *handle, uint32_t timeout)
 	}
 
 	while (lock->state != _UNLOCKED) {
-		signed long elapsed;
+		unsigned int elapsed;
 
 		spin_unlock_irqrestore(&lock->lock, irqflags);
 
@@ -489,7 +490,7 @@ int genlock_wait(struct genlock_handle *handle, uint32_t timeout)
 			break;
 		}
 
-		ticks = (unsigned long) elapsed;
+		ticks = elapsed;
 	}
 
 done:
